+++ lib/ooxml_parser/pptx_parser/pptx_data/presentation/slide/background/background_properties/stretch.rb
@@ -1,0 +1,11 @@
+module OoxmlParser
+  # Class for parsing `stretch` tag
+  class Stretch < OOXMLDocumentObject
+    # Parse Stretch object
+    # @param _node [Nokogiri::XML:Element] node to parse
+    # @return [Stretch] result of parsing
+    def parse(_node)
+      self
+    end
+  end
+end