+++ spec/presentation/slide/background/properties/blip_fill_spec.rb
@@ -1,0 +1,8 @@
+require 'spec_helper'
+
+describe 'BlipFill' do
+  it 'stretch.pptx' do
+    pptx = OoxmlParser::PptxParser.parse_pptx('spec/presentation/slide/background/properties/blip_fill/stretch.pptx')
+    expect(pptx.slides[0].background.properties.blip_fill.stretch).to be_a(OoxmlParser::Stretch)
+  end
+end